+++ src/library_fs.js
@@ -812,6 +812,9 @@ FS.staticInit();` +
       // do the underlying fs rename
       try {
         old_dir.node_ops.rename(old_node, new_dir, new_name);
+        // update old node (we do this here to avoid each backend 
+        // needing to)
+        old_node.parent = new_dir;
       } catch (e) {
         throw e;
       } finally {